+++ local_tests/list_models.py
@@ -4,7 +4,7 @@ import setupenv
 
 import llmprototyping as llmp
 
-llmp.ollama_discover(host='http://192.168.1.13:11434')
+#llmp.ollama_discover(host='http://192.168.1.13:11434')
 
 print('chat completion models:')
 for model_name in llmp.LLMChatCompletionFactory.available_models: