+++ src/llmprototyping/llm_openai.py
@@ -63,6 +63,10 @@ class LLMOpenAIChatCompletion(LLMChatCompletion):
     def from_dict(cls, data):
         return cls(model_name=cls.cls_model_name, context_size=cls.cls_context_size, api_key=data['api_key'])
 
+class OpenAI_GPT_4o_mini(LLMOpenAIChatCompletion):
+    cls_model_name = 'gpt-4o-mini'
+    cls_context_size = 128000
+
 class OpenAI_GPT_4o(LLMOpenAIChatCompletion):
     cls_model_name = 'gpt-4o'
     cls_context_size = 128000